+++ app/Homplexity.hs
@@ -8,12 +8,14 @@
 {-# LANGUAGE TemplateHaskell       #-}
 {-# LANGUAGE UndecidableInstances  #-}
 {-# LANGUAGE ViewPatterns          #-}
+{-# LANGUAGE CPP                   #-}
 -- | Main module parsing inputs, and running analysis.
 module Main (main) where
 
+import Control.Monad(when)
 import Data.Either
 import Data.Functor
-import Data.List
+import Data.List hiding (head)
 import Data.Monoid
 import Data.Version
 
@@ -34,12 +36,38 @@ import System.IO
 
 import Paths_homplexity(version)
 
+import qualified Prelude as P
+import Prelude hiding (head, id, div)
+#ifdef HTML_OUTPUT
+import qualified Data.ByteString.Lazy as B
+import Text.Blaze.Html
+import Text.Blaze.Html.Renderer.Utf8
+import Text.Blaze.Html4.Strict hiding (map)
+import Text.Blaze.Html4.Strict.Attributes hiding (title, style)
+#endif
+
 -- * Command line flag
+defineFlag "v:version" False "Show version"
+
 defineFlag "severity" Warning (concat ["level of output verbosity (", severityOptions, ")"])
 
+data OutputFormat =
+    Text
+#ifdef HTML_OUTPUT
+  | HTML
+#endif
+  deriving (Read, Show)
+defineEQFlag "format" [| Text :: OutputFormat |]
+#ifdef HTML_OUTPUT
+  "{Text|HTML}"
+#else
+  "{Text}"
+#endif
+  "What format to use"
+
 -- | Report to standard error output.
-report ::  String -> IO ()
-report = hPutStrLn stderr
+report :: String -> IO ()
+report  = hPutStrLn stderr
 
 -- * Recursing directory tree in order to list Haskell source files.
 -- | Find all Haskell source files within a given path.
@@ -79,12 +107,27 @@ concatMapM f = fmap concat . mapM f
 -- * Analysis
 -- | Analyze a set of modules.
 analyzeModule :: Module SrcLoc -> IO ()
-analyzeModule  = putStr
-               . concatMap show
+analyzeModule  = out
                . extract flags_severity
                . mconcat metrics
                . program
                . (:[])
+  where
+    out msgs = case flags_format of
+                    Text -> do
+                      putStr . concatMap show $ msgs
+#ifdef HTML_OUTPUT
+                    HTML -> B.putStr $ renderHtml $ do
+                      html $ do
+                        head $ do
+                          title (string "Homplexity Analysis.")
+                        style $ string style_head
+                        body $ do
+                          toHtml . fmap toHtml $ msgs
+    style_head = ".warning  .severity { color: orange; }" <>
+                 ".debug    .severity { color: grey;   }" <>
+                 ".critical .severity { color: red;    }"
+#endif
 
 -- | Process each separate input file.
 processFile :: [Extension] -> FilePath -> IO Bool
@@ -100,21 +143,24 @@ defineFlag "cabal" "" "Project cabal file"
 -- | This flag exists only to make sure that HFLags work.
 defineFlag "fakeFlag" Info "this flag is fake"
 
+versionString = unwords [showVersion version, "git revision", $(gitHash)]
 
 -- | Parse arguments and either process inputs (if available), or suggest proper usage.
 main :: IO ()
 main = do
-  args <- $initHFlags ("Homplexity " ++ showVersion version ++ " " ++ $(gitHash) ++ " - automatic analysis of Haskell code quality")
+  args <- $initHFlags ("Homplexity " ++ versionString ++ " - automatic analysis of Haskell code quality")
+  when flags_version $
+    report $ unwords ["Version: ", versionString]
+
   if null args
     then do report ("Use Haskell source file or directory as an argument, " ++
                     "or use --help to discover options.")
             exitFailure
     else do exts <- cabalExtensions
             sums <- mapM (processFile exts) =<< concatMapM subTrees args
-            putStrLn $ unwords ["Correctly parsed", show $ length $ filter id sums,
-                                "out of",           show $ length             sums,
-                                "input files."]
-
+            report $ unwords ["Correctly parsed", show $ length $ filter P.id sums,
+                              "out of",           show $ length               sums,
+                              "input files."]
 
 cabalExtensions :: IO [Extension]
 cabalExtensions
@@ -123,4 +169,5 @@ cabalExtensions
       parseCabalFile flags_cabal >>=
         either (\msg -> report (show msg) >> return [])
                (\cabal -> return $ languageExtensions Library cabal)
+