+++ lib/Language/Haskell/Homplexity/Assessment.hs
@@ -18,6 +18,8 @@ import Language.Haskell.Homplexity.CodeFragment
 import Language.Haskell.Homplexity.Cyclomatic
 import Language.Haskell.Homplexity.Message
 import Language.Haskell.Homplexity.Metric
+import Language.Haskell.Homplexity.RecordFieldsCount
+import Language.Haskell.Homplexity.TypeClassComplexity
 import Language.Haskell.Homplexity.TypeComplexity
 
 import HFlags
@@ -76,11 +78,11 @@ defineFlag "moduleLinesCritical" (3000 :: Int) "issue critical when module excee
 
 assessModuleLength :: Assessment LOC
 assessModuleLength (fromIntegral -> locs)
-                   | locs > flags_moduleLinesWarning  = (Warning,  "should be kept below "        ++
-                                                                    show flags_moduleLinesWarning ++
-                                                                   " lines of code.")
                    | locs > flags_moduleLinesCritical = (Critical, "this function exceeds "       ++
                                                                     show flags_moduleLinesCritical ++
+                                                                   " lines of code.")
+                   | locs > flags_moduleLinesWarning  = (Warning,  "should be kept below "        ++
+                                                                    show flags_moduleLinesWarning ++
                                                                    " lines of code.")
                    | otherwise    = (Info,     ""                                        )
 
@@ -91,11 +93,11 @@ defineFlag "functionLinesCritical" (40 :: Int) "issue critical when function exc
 
 assessFunctionLength :: Assessment LOC
 assessFunctionLength (fromIntegral -> locs)
-                   | locs > flags_functionLinesWarning  = (Warning,  "should be kept below "          ++
-                                                                       show flags_functionLinesWarning ++
-                                                                      " lines of code.")
                    | locs > flags_functionLinesCritical = (Critical, "this function exceeds "          ++
                                                                        show flags_functionLinesCritical ++
+                                                                      " lines of code.")
+                   | locs > flags_functionLinesWarning  = (Warning,  "should be kept below "          ++
+                                                                       show flags_functionLinesWarning ++
                                                                       " lines of code.")
                    | otherwise                          = (Info,     ""                                 )
 
@@ -106,12 +108,12 @@ defineFlag "functionDepthCritical" (8 :: Int) "issue critical when function exce
 
 assessFunctionDepth :: Assessment Depth
 assessFunctionDepth (fromIntegral -> depth)
-                    | depth > flags_functionDepthWarning = (Warning, "should have no more than " ++
-                                                                      show depth                 ++
-                                                                     " nested conditionals"            )
-                    | depth > flags_functionDepthWarning = (Warning, "should never exceed " ++
-                                                                      show depth            ++
-                                                                     " nesting levels for conditionals")
+                    | depth > flags_functionDepthCritical = (Critical, "should never exceed " ++
+                                                                        show depth            ++
+                                                                       " nesting levels for conditionals")
+                    | depth > flags_functionDepthWarning  = (Warning,  "should have no more than " ++
+                                                                        show depth                 ++
+                                                                       " nested conditionals"            )
                     | otherwise = (Info,    ""                                )
 
 -- *** Cyclomatic complexity of function definition
@@ -120,11 +122,11 @@ defineFlag "functionCCCritical" (50::Int) "issue critical when function's cyclom
 
 assessFunctionCC :: Assessment Cyclomatic
 assessFunctionCC (fromIntegral -> cy)
-                 | cy > flags_functionCCWarning  = (Warning, "should be less than "        ++
-                                                              show flags_functionCCWarning)
-                 | cy > flags_functionCCCritical = (Warning, "must never be as high as " ++
-                                                              show flags_functionCCCritical)
-                 | otherwise                     = (Info,    ""                               )
+                 | cy > flags_functionCCCritical = (Critical, "must never be as high as " ++
+                                                               show flags_functionCCCritical)
+                 | cy > flags_functionCCWarning  = (Warning,  "should be less than "        ++
+                                                               show flags_functionCCWarning)
+                 | otherwise                     = (Info,     ""                               )
 
 -- ** Type signature complexity
 -- *** Type constructor depth in each type signature
@@ -133,11 +135,11 @@ defineFlag "typeConDepthCritical" (9::Int) "issue critical when type constructor
 
 assessTypeConDepth :: Assessment ConDepth
 assessTypeConDepth (fromIntegral -> cy)
-                 | cy > flags_typeConDepthWarning  = (Warning, "should be less than "        ++
-                                                                show flags_typeConDepthWarning )
-                 | cy > flags_typeConDepthCritical = (Warning, "must never be as high as " ++
-                                                                show flags_typeConDepthCritical)
-                 | otherwise                       = (Info,    ""                              )
+                 | cy > flags_typeConDepthCritical = (Critical, "must never be as high as " ++
+                                                                 show flags_typeConDepthCritical)
+                 | cy > flags_typeConDepthWarning  = (Warning,  "should be less than "        ++
+                                                                 show flags_typeConDepthWarning )
+                 | otherwise                       = (Info,     ""                              )
 
 -- *** Number of function arguments mentioned in each type signature
 defineFlag "numFunArgsWarning"  (5::Int) "issue warning when number of function arguments exceeds this number"
@@ -145,17 +147,57 @@ defineFlag "numFunArgsCritical" (9::Int) "issue critical when number of function
 
 assessNumFunArgs :: Assessment NumFunArgs
 assessNumFunArgs (fromIntegral -> cy)
-                 | cy > flags_numFunArgsWarning  = (Warning, "should be less than " ++ show flags_numFunArgsWarning )
-                 | cy > flags_numFunArgsCritical = (Warning, "must never reach "    ++ show flags_numFunArgsCritical)
-                 | otherwise                     = (Info,    ""                                                     )
+                 | cy > flags_numFunArgsCritical = (Critical, "must never reach "    ++ show flags_numFunArgsCritical)
+                 | cy > flags_numFunArgsWarning  = (Warning,  "should be less than " ++ show flags_numFunArgsWarning )
+                 | otherwise                     = (Info,     ""                                                     )
+
+-- ** Data type complexity
+-- *** Record fields count
+defineFlag "recordFieldsCountWarning"  (6::Int) "issue warning when combined record fields count exceeds this number"
+defineFlag "recordFieldsCountCritical" (9::Int) "issue critical when combined record fields count exceeds this number"
+
+assessRecordFieldsCount :: Assessment RecordFieldsCount
+assessRecordFieldsCount (fromIntegral -> cy)
+                        | cy > flags_recordFieldsCountCritical = (Critical, "must never reach " ++ show flags_recordFieldsCountCritical  )
+                        | cy > flags_recordFieldsCountWarning  = (Warning,  "should be less than " ++ show flags_recordFieldsCountWarning)
+                        | otherwise                            = (Info,     ""                                                           )
+
+-- ** Type class complexity
+-- *** Method count of type class
+defineFlag "typeClassNonTypeDeclWarning"  (5::Int) "issue warning when the number of methods in a type class exceeds this number"
+defineFlag "typeClassNonTypeDeclCritical" (7::Int) "issue critical when the number of methods in a type class exceeds this number"
+
+assessTCNonTypeDeclCount :: Assessment NonTypeDeclCount
+assessTCNonTypeDeclCount (fromIntegral -> mc)
+  | mc > flags_typeClassNonTypeDeclCritical = (Critical, "should never have more than " ++
+                                              show flags_typeClassNonTypeDeclCritical)
+  | mc > flags_typeClassNonTypeDeclWarning  = (Warning,  " should have no more than " ++
+                                              show flags_typeClassNonTypeDeclWarning)
+  | otherwise = (Info, "")
+
+-- *** Associated type count of type class
+defineFlag "typeClassAssocTypesWarning"  (3::Int) "issue warning when the number of associated types in a type class exceeds this number"
+defineFlag "typeClassAssocTypesCritical" (5::Int) "issue critical when the number of associated types in a type class exceeds this number"
+
+assessTCAssocTypesCount :: Assessment AssocTypeCount
+assessTCAssocTypesCount (fromIntegral -> atc)
+  | atc > flags_typeClassAssocTypesCritical = (Critical, "should never have more than " ++
+                                              show flags_typeClassAssocTypesCritical)
+  | atc > flags_typeClassAssocTypesWarning  = (Warning,  " should have no more than " ++
+                                              show flags_typeClassAssocTypesWarning)
+  | otherwise                               = (Info, "")
 
 -- * Computing and assessing @Metric@s for all @CodeFragment@.
 -- | Compute all metrics, and assign severity depending on configured thresholds.
 metrics :: [Program -> Log]
-metrics  = [measureTopOccurs assessModuleLength   locT        moduleT
-           ,measureTopOccurs assessFunctionLength locT        functionT
-           ,measureTopOccurs assessFunctionDepth  depthT      functionT
-           ,measureTopOccurs assessFunctionCC     cyclomaticT functionT
-           ,measureTopOccurs assessTypeConDepth   conDepthT   typeSignatureT
-           ,measureTopOccurs assessNumFunArgs     numFunArgsT typeSignatureT]
+metrics  = [ measureTopOccurs assessModuleLength       locT               moduleT
+           , measureTopOccurs assessFunctionLength     locT               functionT
+           , measureTopOccurs assessFunctionDepth      depthT             functionT
+           , measureTopOccurs assessFunctionCC         cyclomaticT        functionT
+           , measureTopOccurs assessTypeConDepth       conDepthT          typeSignatureT
+           , measureTopOccurs assessNumFunArgs         numFunArgsT        typeSignatureT
+           , measureTopOccurs assessRecordFieldsCount  recordFieldsCountT dataDefT
+           , measureTopOccurs assessTCNonTypeDeclCount nonTypeDeclCountT  typeClassT
+           , measureTopOccurs assessTCAssocTypesCount  assocTypeCountT    typeClassT
+           ]