+++ lib/Language/Haskell/Homplexity/Message.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE RecordWildCards            #-}
 {-# LANGUAGE TemplateHaskell            #-}
+{-# LANGUAGE OverloadedStrings          #-}
 -- | Classifying messages by severity and filtering them.
 module Language.Haskell.Homplexity.Message (
     Log
@@ -26,9 +27,15 @@ import Data.Monoid
 import Data.Semigroup
 #endif
 import Data.Sequence                            as Seq
-import Language.Haskell.Exts
+import Language.Haskell.Exts hiding (style)
 import Language.Haskell.TH.Syntax                           (Lift(..))
 import HFlags
+#ifdef HTML_OUTPUT
+import Prelude hiding (head, id, div, span)
+import Text.Blaze.Html4.Strict hiding (map, style)
+import Text.Blaze.Html4.Strict.Attributes hiding (title, span)
+import Text.Blaze.Renderer.Utf8 (renderMarkup)
+#endif
 
 -- | Keeps a set of messages
 newtype Log = Log { unLog :: Seq Message }
@@ -62,9 +69,36 @@ instance Show Message where
                                                   . shows loc
                                                   -- . shows srcLine
                                                   -- . shows srcColumn
-                                                  . (':':)
+                                                  . (": "++)
                                                   . (msgText++)
                                                   . ('\n':)
+
+#ifdef HTML_OUTPUT
+instance ToMarkup Message where
+  toMarkup Message {..} =
+    p ! classId $
+     (toMarkup msgSeverity
+       <> string ": "
+       <> toMarkup msgSrc
+       <> string ": "
+       <> string msgText)
+    where
+      classId = case msgSeverity of
+                     Debug    -> class_ "debug"
+                     Info     -> class_ "info"
+                     Warning  -> class_ "warning"
+                     Critical -> class_ "critical"
+
+instance ToMarkup Severity where
+  toMarkup Debug    = span   ! class_ "severity" $ string (show Debug)
+  toMarkup Info     = span   ! class_ "severity" $ string (show Info)
+  toMarkup Warning  = strong ! class_ "severity" $ string (show Warning)
+  toMarkup Critical = strong ! class_ "severity" $ string (show Critical)
+
+instance ToMarkup SrcLoc where
+  toMarkup SrcLoc {..} = a ! href (toValue srcFilename) $ (string srcFilename)
+
+#endif
 
 -- | Message severity
 data Severity = Debug
@@ -93,7 +127,7 @@ instance FlagType Severity where
 message ::  Severity -> SrcLoc -> String -> Log
 message msgSeverity msgSrc msgText = Log $ Seq.singleton Message {..}
 
--- | TODO: automatic inference of the srcLine 
+-- | TODO: automatic inference of the srcLine
 -- | Log a certain error
 critical :: SrcLoc -> String -> Log
 critical  = message Critical